Split Root into provider/controller
